+++ src/main/java/com/example/bigbrotherbe/global/email/entity/Email.java
@@ -1,0 +1,9 @@
+package com.example.bigbrotherbe.global.email.entity;
+
+
+import lombok.Builder;
+
+@Builder
+public record Email(String title, String authCode, String toEmailAddress) {
+
+}
