+++ src/main/java/com/example/bigbrotherbe/global/email/entity/EmailRequest.java
@@ -1,4 +1,4 @@
-package com.example.bigbrotherbe.global.email;
+package com.example.bigbrotherbe.global.email.entity;
 
 import lombok.Getter;
