+++ src/main/java/com/example/bigbrotherbe/global/email/entity/EmailVerificationResult.java
@@ -1,11 +1,10 @@
-package com.example.bigbrotherbe.global.email;
+package com.example.bigbrotherbe.global.email.entity;
 
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 @Getter
 @Builder
